Modif to wettying/Drying treatment with insertion/removal:
1/ two top layer nodal indexes: ilhkv (nodal-continuous), ilhev
(elemental-discontinuous). ilhev is used in dry element where the
free-surface is discontinuous. ilhkv is used in nodal loop and for w/d
nodes is computed only on the wet side.
2/ nodal depth hdknv is computed, as it was in the master, from all
nodes: this helps when computing bottom vertical velocities and make the
code more robust. However conservativeness is spoiled for w/d nodes.